// server/src/types/media.d.ts
export interface MediaAttributes {
	id: number
	/**媒体类型 */
	type: string
	/**媒体地址 */
	src: string
	/**描述 */
	alt: string
	/**宽 */
	width: number
	/**高 */
	height: number
	/**变体 ID */
	variant_ids: string
}
export interface ImageAttributes extends MediaAttributes {}
export interface VideoAttributes extends MediaAttributes {
	preview_image_id?: number
}